refactor(brain-dump): extract input validation and plan parse helpers
Port the leftover from #126 that is not auth or beta gating. Same
messages and limits as the inlined prioritize path; now unit-tested
without a Convex action.

Skipped #126 betaGating, userResolver, RevenueCat, and requireUser
rewrites.

Task leftover: PR #126
